refactor: move resources::flavor_group types from lib to wire crate
