+++ lib/src/resources/flavor_group.rs
@@ -1,79 +1,15 @@
 use crate::common::{request, request_bare, SerializableNone};
 use crate::error::ApiError;
-use crate::resources::FlavorMinimal;
-use crate::user::ProjectMinimal;
 use anyhow::Context;
+use lrzcc_wire::resources::{
+    FlavorGroup, FlavorGroupCreateData, FlavorGroupCreated,
+    FlavorGroupDetailed, FlavorGroupInitialize, FlavorGroupModifyData,
+    FlavorGroupUsage, FlavorGroupUsageAggregate,
+};
 use reqwest::blocking::Client;
 use reqwest::Url;
 use reqwest::{Method, StatusCode};
-use serde::{Deserialize, Serialize};
-use std::fmt::Display;
 use std::rc::Rc;
-use tabled::Tabled;
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorGroup {
-    pub id: u32,
-    pub name: String,
-    #[tabled(skip)]
-    pub flavors: Vec<u32>,
-    pub project: u32,
-}
-
-impl Display for FlavorGroup {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("FlavorGroup(id={}, name={})", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorGroupMinimal {
-    pub id: u32,
-    pub name: String,
-}
-
-// TODO maybe rethink the Display implementations
-impl Display for FlavorGroupMinimal {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("FlavorGroup(id={}, name={})", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorGroupDetailed {
-    pub id: u32,
-    pub name: String,
-    #[tabled(skip)]
-    pub flavors: Vec<FlavorMinimal>,
-    pub project: ProjectMinimal,
-}
-
-impl Display for FlavorGroupDetailed {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("FlavorGroup(id={}, name={})", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorGroupCreated {
-    pub id: u32,
-    pub name: String,
-    #[tabled(skip)]
-    pub flavors: Vec<FlavorMinimal>,
-    pub project: u32,
-}
-
-impl Display for FlavorGroupCreated {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("FlavorGroup(id={}, name={})", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorGroupInitialize {
-    pub new_flavor_group_count: u32,
-    pub new_flavor_count: u32,
-}
 
 pub struct FlavorGroupApi {
     pub url: String,
@@ -123,21 +59,6 @@ impl FlavorGroupListRequest {
     }
 }
 
-#[derive(Clone, Debug, Serialize)]
-struct FlavorGroupCreateData {
-    name: String,
-    flavors: Vec<u32>,
-}
-
-impl FlavorGroupCreateData {
-    fn new(name: String) -> Self {
-        Self {
-            name,
-            flavors: vec![],
-        }
-    }
-}
-
 pub struct FlavorGroupCreateRequest {
     url: String,
     client: Rc<Client>,
@@ -162,26 +83,6 @@ impl FlavorGroupCreateRequest {
             Some(&self.data),
             StatusCode::CREATED,
         )
-    }
-}
-
-#[derive(Clone, Debug, Serialize)]
-struct FlavorGroupModifyData {
-    id: u32,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    project: Option<u32>,
-}
-
-impl FlavorGroupModifyData {
-    fn new(id: u32) -> Self {
-        Self {
-            id,
-            name: None,
-            project: None,
-        }
     }
 }
 
@@ -220,22 +121,6 @@ impl FlavorGroupModifyRequest {
             StatusCode::OK,
         )
     }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorGroupUsage {
-    pub user_id: u32,
-    pub user_name: String,
-    pub flavorgroup_id: u32,
-    pub flavorgroup_name: String,
-    pub usage: u32,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorGroupUsageAggregate {
-    pub flavorgroup_id: u32,
-    pub flavorgroup_name: String,
-    pub usage: u32,
 }
 
 pub struct FlavorGroupUsageRequest {
